requirements.txt
jsonpickle==2.1.0
requests==2.27.1
setuptools==58.1.0
python-dotenv==0.20.0
faker == 14.2.0